Oppdatert delt bosted validering
